Change decorator name to "verify_linearizability". Move the decorator into skvbc_history_tracker.py. Remove unnecessary imports from test_skvbc_linearizability.py.
